Move shield VEF patch to VEFcompat.cs
